+++ CHANGELOG.md
@@ -15,8 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * PhysioNet 2012 challenge time stamp is now in hours (previously minutes) for consistency with the PhysioNet 2019 challenge
 * PhysioNet 2012 challenge time stamp is now time since admission, not row index
-* Impute all channels
-* Standardise all channels
+* Impute time series channels only
+* Standardise continuous channels only
 
 ### Fixed